photography_agents/requirements.txt
google-adk>=0.1.0
google-genai>=0.1.0
pillow
python-dotenv
